Add workaround for ending simulation early
a68436f6d87828f99ad2b4411daf28f93b377470 introduced a change to enable
the usage of KVM. That commit in itself is just a workaround to properly
run the simulation and leads to the following problem:

When a program uses a `wfi` instruction and is utilizing KVM as well,
it's possible for a simulation to end early, because SystemC thinks
there's no more work to do, even though KVM is still waiting for an
interrupt. This is usually not a problem for QQVP, because there's lots
work happening. Some tests (simple-write-test and dmi-test) on the other
hand don't do much. That's why sometimes it does happen, that the
simulation ends before all (or any!) work is done.

By adding a `keep_alive` variable it's possible to control when the
simulation ends.

But keep in mind: THIS IS JUST A WORKAROUND (!!!) until a better
solution for dealing with KVM is found.

Signed-off-by: Marco Liebel <quic_mliebel@quicinc.com>